Move dkron command to it's own package
Better code org
