+++ dkron/agent.go
@@ -7,12 +7,9 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net"
-	"os"
-	"os/signal"
 	"runtime"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/Sirupsen/logrus"
@@ -20,13 +17,9 @@ import (
 	metrics "github.com/armon/go-metrics"
 	"github.com/hashicorp/memberlist"
 	"github.com/hashicorp/serf/serf"
-	"github.com/mitchellh/cli"
 )
 
 const (
-	// gracefulTimeout controls how long we wait before forcefully terminating
-	gracefulTimeout = 3 * time.Second
-
 	defaultRecoverTime = 10 * time.Second
 )
 
@@ -39,15 +32,8 @@ var (
 	defaultLeaderTTL = 20 * time.Second
 )
 
-// ProcessorFactory is a function type that creates a new instance
-// of a processor.
-type ProcessorFactory func() (ExecutionProcessor, error)
-
-// AgentCommand run server
-type AgentCommand struct {
-	Ui               cli.Ui
+type Agent struct {
 	Version          string
-	ShutdownCh       <-chan struct{}
 	ProcessorPlugins map[string]ExecutionProcessor
 	ExecutorPlugins  map[string]Executor
 	HTTPTransport    Transport
@@ -61,79 +47,55 @@ type AgentCommand struct {
 	ready     bool
 }
 
-// Help returns agent command usage to the CLI.
-func (a *AgentCommand) Help() string {
-	helpText := `
-Usage: dkron agent [options]
-	Run dkron agent
+// ProcessorFactory is a function type that creates a new instance
+// of a processor.
+type ProcessorFactory func() (ExecutionProcessor, error)
 
-Options:
+func Create(config *Config) (*Agent, error) {
+	a := &Agent{
+		config: config,
+	}
 
-  -bind-addr=0.0.0.0:8946         Address to bind network listeners to.
-  -advertise-addr=bind_addr       Address used to advertise to other nodes in the cluster. By default, the bind address is advertised.
-  -http-addr=0.0.0.0:8080         Address to bind the UI web server to. Only used when server.
-  -discover=cluster               A cluster name used to discovery peers. On
-                                  networks that support multicast, this can be used to have
-                                  peers join each other without an explicit join.
-  -join=addr                      An initial agent to join with. This flag can be
-                                  specified multiple times.
-  -node=hostname                  Name of this node. Must be unique in the cluster
-  -profile=[lan|wan|local]        Profile is used to control the timing profiles used.
-                                  The default if not provided is lan.
-  -server=false                   This node is running in server mode.
-  -tag key=value                  Tag can be specified multiple times to attach multiple
-                                  key/value tag pairs to the given node.
-  -keyspace=dkron                 The keyspace to use. A prefix under all data is stored
-                                  for this instance.
-  -backend=[etcd|consul|zk]       Backend storage to use, etcd, consul or zookeeper. The default
-                                  is etcd.
-  -backend-machine=127.0.0.1:2379 Backend storage servers addresses to connect to. This flag can be
-                                  specified multiple times.
-  -encrypt                        Key for encrypting network traffic.
-                                  Must be a base64-encoded 16-byte key.
-  -rpc-port=6868                  RPC Port used to communicate with clients. Only used when server.
-                                  The RPC IP Address will be the same as the bind address.
-  -advertise-rpc-port             Use the value of -rpc-port by default
+	s, err := a.setupSerf()
+	if err != nil {
+		return nil, fmt.Errorf("agent: Can not setup serf, %s", err)
+	}
+	a.serf = s
+	a.join(config.StartJoin, true)
 
-  -mail-host                      Mail server host address to use for notifications.
-  -mail-port                      Mail server port.
-  -mail-username                  Mail server username used for authentication.
-  -mail-password                  Mail server password to use.
-  -mail-from                      From email address to use.
+	if err := initMetrics(a); err != nil {
+		log.Fatal("agent: Can not setup metrics")
+	}
 
-  -webhook-url                    Webhook url to call for notifications.
-  -webhook-payload                Body of the POST request to send on webhook call.
-  -webhook-header                 Headers to use when calling the webhook URL. Can be specified multiple times.
+	// Expose the node name
+	expNode.Set(a.config.NodeName)
 
-  -log-level=info                 Log level (debug, info, warn, error, fatal, panic). Default to info.
+	if config.Server {
+		a.StartServer()
+	}
+	go a.eventLoop()
+	a.ready = true
 
-  -dog-statsd-addr                DataDog Agent address
-  -dog-statsd-tags                Datadog tags, specified as key:value
-  -statsd-addr                    Statsd Address
-`
-	return strings.TrimSpace(helpText)
+	return a, nil
 }
 
 // setupSerf is used to create the agent we use
-func (a *AgentCommand) setupSerf() *serf.Serf {
+func (a *Agent) setupSerf() (*serf.Serf, error) {
 	config := a.config
 
 	bindIP, bindPort, err := config.AddrParts(config.BindAddr)
 	if err != nil {
-		a.Ui.Error(fmt.Sprintf("Invalid bind address: %s", err))
-		return nil
+		return nil, fmt.Errorf("Invalid bind address: %s", err)
 	}
 
 	// Check if we have an interface
 	if iface, _ := config.NetworkInterface(); iface != nil {
 		addrs, err := iface.Addrs()
 		if err != nil {
-			a.Ui.Error(fmt.Sprintf("Failed to get interface addresses: %s", err))
-			return nil
+			return nil, fmt.Errorf("Failed to get interface addresses: %s", err)
 		}
 		if len(addrs) == 0 {
-			a.Ui.Error(fmt.Sprintf("Interface '%s' has no addresses", config.Interface))
-			return nil
+			return nil, fmt.Errorf("Interface '%s' has no addresses", config.Interface)
 		}
 
 		// If there is no bind IP, pick an address
@@ -164,8 +126,7 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 				// Found an IP
 				found = true
 				bindIP = addrIP.String()
-				a.Ui.Output(fmt.Sprintf("Using interface '%s' address '%s'",
-					config.Interface, bindIP))
+				log.Infof("Using interface '%s' address '%s'", config.Interface, bindIP)
 
 				// Update the configuration
 				bindAddr := &net.TCPAddr{
@@ -176,8 +137,7 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 				break
 			}
 			if !found {
-				a.Ui.Error(fmt.Sprintf("Failed to find usable address for interface '%s'", config.Interface))
-				return nil
+				return nil, fmt.Errorf("Failed to find usable address for interface '%s'", config.Interface)
 			}
 
 		} else {
@@ -194,9 +154,8 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 				}
 			}
 			if !found {
-				a.Ui.Error(fmt.Sprintf("Interface '%s' has no '%s' address",
-					config.Interface, bindIP))
-				return nil
+				return nil, fmt.Errorf("Interface '%s' has no '%s' address",
+					config.Interface, bindIP)
 			}
 		}
 	}
@@ -206,8 +165,7 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 	if config.AdvertiseAddr != "" {
 		advertiseIP, advertisePort, err = config.AddrParts(config.AdvertiseAddr)
 		if err != nil {
-			a.Ui.Error(fmt.Sprintf("Invalid advertise address: %s", err))
-			return nil
+			return nil, fmt.Errorf("Invalid advertise address: %s", err)
 		}
 	}
 	//Ues the value of "RPCPort" if AdvertiseRPCPort has not been set
@@ -217,8 +175,7 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 
 	encryptKey, err := config.EncryptBytes()
 	if err != nil {
-		a.Ui.Error(fmt.Sprintf("Invalid encryption key: %s", err))
-		return nil
+		return nil, fmt.Errorf("Invalid encryption key: %s", err)
 	}
 
 	serfConfig := serf.DefaultConfig()
@@ -230,8 +187,7 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 	case "local":
 		serfConfig.MemberlistConfig = memberlist.DefaultLocalConfig()
 	default:
-		a.Ui.Error(fmt.Sprintf("Unknown profile: %s", config.Profile))
-		return nil
+		return nil, fmt.Errorf("Unknown profile: %s", config.Profile)
 	}
 
 	serfConfig.MemberlistConfig.BindAddr = bindIP
@@ -266,7 +222,6 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 	serfConfig.EventCh = a.eventCh
 
 	// Start Serf
-	a.Ui.Output("Starting Dkron agent...")
 	log.Info("agent: Dkron agent starting")
 
 	serfConfig.LogOutput = ioutil.Discard
@@ -275,16 +230,15 @@ func (a *AgentCommand) setupSerf() *serf.Serf {
 	// Create serf first
 	serf, err := serf.Create(serfConfig)
 	if err != nil {
-		a.Ui.Error(err.Error())
 		log.Error(err)
-		return nil
+		return nil, err
 	}
 
-	return serf
+	return serf, nil
 }
 
 // Config returns the agent's config.
-func (a *AgentCommand) Config() *Config {
+func (a *Agent) Config() *Config {
 	return a.config
 }
 
@@ -302,33 +256,7 @@ func UnmarshalTags(tags []string) (map[string]string, error) {
 	return result, nil
 }
 
-// Run will execute the main functions of the agent command.
-// This includes the main eventloop and starting the server if enabled.
-//
-// The returned value is the exit code.
-func (a *AgentCommand) Run(args []string) int {
-	a.config = NewConfig(args, a.Version)
-	if a.serf = a.setupSerf(); a.serf == nil {
-		log.Fatal("agent: Can not setup serf")
-	}
-	a.join(a.config.StartJoin, true)
-
-	if i := initMetrics(a); i != 0 {
-		return i
-	}
-
-	// Expose the node name
-	expNode.Set(a.config.NodeName)
-
-	if a.config.Server {
-		a.StartServer()
-	}
-	go a.eventLoop()
-	a.ready = true
-	return a.handleSignals()
-}
-
-func (a *AgentCommand) StartServer() {
+func (a *Agent) StartServer() {
 	if a.Store == nil {
 		a.Store = NewStore(a.config.Backend, a.config.BackendMachines, a, a.config.Keyspace, nil)
 	}
@@ -343,84 +271,7 @@ func (a *AgentCommand) StartServer() {
 	a.participate()
 }
 
-// handleSignals blocks until we get an exit-causing signal
-func (a *AgentCommand) handleSignals() int {
-	signalCh := make(chan os.Signal, 4)
-	signal.Notify(signalCh, os.Interrupt, syscall.SIGTERM, syscall.SIGHUP)
-
-WAIT:
-	// Wait for a signal
-	var sig os.Signal
-	select {
-	case s := <-signalCh:
-		sig = s
-	case <-a.ShutdownCh:
-		sig = os.Interrupt
-	}
-	a.Ui.Output(fmt.Sprintf("Caught signal: %v", sig))
-
-	// Check if this is a SIGHUP
-	if sig == syscall.SIGHUP {
-		a.handleReload()
-		goto WAIT
-	}
-
-	// Check if we should do a graceful leave
-	graceful := false
-	if sig == syscall.SIGTERM || sig == os.Interrupt {
-		graceful = true
-	}
-
-	// Fail fast if not doing a graceful leave
-	if !graceful {
-		return 1
-	}
-
-	// Attempt a graceful leave
-	gracefulCh := make(chan struct{})
-	a.Ui.Output("Gracefully shutting down agent...")
-	log.Info("agent: Gracefully shutting down agent...")
-	go func() {
-		if err := a.serf.Leave(); err != nil {
-			a.Ui.Error(fmt.Sprintf("Error: %s", err))
-			log.Error(fmt.Sprintf("Error: %s", err))
-			return
-		}
-		close(gracefulCh)
-	}()
-
-	// Wait for leave or another signal
-	select {
-	case <-signalCh:
-		return 1
-	case <-time.After(gracefulTimeout):
-		return 1
-	case <-gracefulCh:
-		return 0
-	}
-}
-
-// handleReload is invoked when we should reload our configs, e.g. SIGHUP
-func (a *AgentCommand) handleReload() {
-	a.Ui.Output("Reloading configuration...")
-	newConf := readConfig(a.Version)
-	if newConf == nil {
-		a.Ui.Error(fmt.Sprintf("Failed to reload configs"))
-		return
-	}
-	a.config = newConf
-
-	// Reset serf tags
-	a.serf.SetTags(a.config.Tags)
-	//Config reloading will also reload Notification settings
-}
-
-// Synopsis returns the purpose of the command for the CLI
-func (a *AgentCommand) Synopsis() string {
-	return "Run dkron"
-}
-
-func (a *AgentCommand) participate() {
+func (a *Agent) participate() {
 	a.candidate = leadership.NewCandidate(a.Store.Client, a.Store.LeaderKey(), a.config.NodeName, defaultLeaderTTL)
 
 	go func() {
@@ -433,7 +284,7 @@ func (a *AgentCommand) participate() {
 }
 
 // Leader election routine
-func (a *AgentCommand) runForElection() {
+func (a *Agent) runForElection() {
 	log.Info("agent: Running for election")
 	defer metrics.MeasureSince([]string{"agent", "runForElection"}, time.Now())
 	electedCh, errCh := a.candidate.RunForElection()
@@ -464,7 +315,7 @@ func (a *AgentCommand) runForElection() {
 }
 
 // Utility method to get leader nodename
-func (a *AgentCommand) leaderMember() (*serf.Member, error) {
+func (a *Agent) leaderMember() (*serf.Member, error) {
 	leaderName := a.Store.GetLeader()
 	for _, member := range a.serf.Members() {
 		if member.Name == string(leaderName) {
@@ -474,7 +325,7 @@ func (a *AgentCommand) leaderMember() (*serf.Member, error) {
 	return nil, ErrLeaderNotFound
 }
 
-func (a *AgentCommand) listServers() []serf.Member {
+func (a *Agent) listServers() []serf.Member {
 	members := []serf.Member{}
 
 	for _, member := range a.serf.Members() {
@@ -489,13 +340,13 @@ func (a *AgentCommand) listServers() []serf.Member {
 
 // GetBindIP returns the IP address that the agent is bound to.
 // This could be different than the originally configured address.
-func (a *AgentCommand) GetBindIP() (string, error) {
+func (a *Agent) GetBindIP() (string, error) {
 	bindIP, _, err := a.config.AddrParts(a.config.BindAddr)
 	return bindIP, err
 }
 
 // Listens to events from Serf and handle the event.
-func (a *AgentCommand) eventLoop() {
+func (a *Agent) eventLoop() {
 	serfShutdownCh := a.serf.ShutdownCh()
 	log.Info("agent: Listen for events")
 	for {
@@ -583,7 +434,7 @@ func (a *AgentCommand) eventLoop() {
 }
 
 // Start or restart scheduler
-func (a *AgentCommand) schedule() {
+func (a *Agent) schedule() {
 	log.Debug("agent: Restarting scheduler")
 	jobs, err := a.Store.GetJobs()
 	if err != nil {
@@ -593,7 +444,7 @@ func (a *AgentCommand) schedule() {
 }
 
 // Join asks the Serf instance to join. See the Serf.Join function.
-func (a *AgentCommand) join(addrs []string, replay bool) (n int, err error) {
+func (a *Agent) join(addrs []string, replay bool) (n int, err error) {
 	log.Infof("agent: joining: %v replay: %v", addrs, replay)
 	n, err = a.serf.Join(addrs, !replay)
 	if n > 0 {
@@ -605,7 +456,7 @@ func (a *AgentCommand) join(addrs []string, replay bool) (n int, err error) {
 	return
 }
 
-func (a *AgentCommand) processFilteredNodes(job *Job) ([]string, map[string]string, error) {
+func (a *Agent) processFilteredNodes(job *Job) ([]string, map[string]string, error) {
 	var nodes []string
 	tags := make(map[string]string)
 
@@ -644,7 +495,7 @@ func (a *AgentCommand) processFilteredNodes(job *Job) ([]string, map[string]stri
 	return nodes, tags, nil
 }
 
-func (a *AgentCommand) setExecution(payload []byte) *Execution {
+func (a *Agent) setExecution(payload []byte) *Execution {
 	var ex Execution
 	if err := json.Unmarshal(payload, &ex); err != nil {
 		log.Fatal(err)
@@ -661,7 +512,7 @@ func (a *AgentCommand) setExecution(payload []byte) *Execution {
 // This function is called when a client request the RPCAddress
 // of the current member.
 // in marathon, it would return the host's IP and advertise RPC port
-func (a *AgentCommand) getRPCAddr() string {
+func (a *Agent) getRPCAddr() string {
 	bindIP := a.serf.LocalMember().Addr
 
 	return fmt.Sprintf("%s:%d", bindIP, a.config.AdvertiseRPCPort)
