+++ dkron/dashboard.go
@@ -29,7 +29,7 @@ type commonDashboardData struct {
 	Keyspace   string
 }
 
-func newCommonDashboardData(a *AgentCommand, nodeName, path string) *commonDashboardData {
+func newCommonDashboardData(a *Agent, nodeName, path string) *commonDashboardData {
 	leaderName := ""
 	l, err := a.leaderMember()
 	if err == nil {
@@ -49,7 +49,7 @@ func newCommonDashboardData(a *AgentCommand, nodeName, path string) *commonDashb
 }
 
 // dashboardRoutes registers dashboard specific routes on the gin RouterGroup.
-func (a *AgentCommand) dashboardRoutes(r *gin.RouterGroup) {
+func (a *Agent) dashboardRoutes(r *gin.RouterGroup) {
 	r.GET("/static/*asset", servePublic)
 
 	dashboard := r.Group("/" + dashboardPathPrefix)
@@ -58,7 +58,7 @@ func (a *AgentCommand) dashboardRoutes(r *gin.RouterGroup) {
 	dashboard.GET("/jobs/:job/executions", a.dashboardExecutionsHandler)
 }
 
-func (a *AgentCommand) dashboardIndexHandler(c *gin.Context) {
+func (a *Agent) dashboardIndexHandler(c *gin.Context) {
 	data := struct {
 		Common *commonDashboardData
 	}{
@@ -67,7 +67,7 @@ func (a *AgentCommand) dashboardIndexHandler(c *gin.Context) {
 	c.HTML(http.StatusOK, "index", data)
 }
 
-func (a *AgentCommand) dashboardJobsHandler(c *gin.Context) {
+func (a *Agent) dashboardJobsHandler(c *gin.Context) {
 	jobs, _ := a.Store.GetJobs()
 
 	data := struct {
@@ -81,7 +81,7 @@ func (a *AgentCommand) dashboardJobsHandler(c *gin.Context) {
 	c.HTML(http.StatusOK, "jobs", data)
 }
 
-func (a *AgentCommand) dashboardExecutionsHandler(c *gin.Context) {
+func (a *Agent) dashboardExecutionsHandler(c *gin.Context) {
 	job := c.Param("job")
 
 	groups, byGroup, err := a.Store.GetGroupedExecutions(job)
